requirements.txt
fastapi==0.115.6
uvicorn==0.32.0
SQLAlchemy==2.0.36
fastapi-utils==0.7.0
typing-inspect==0.9.0
certifi==2024.8.30
pycparser==2.22
fastapi-users==14.0.0
Authlib==1.3.2
psycopg2-binary==2.9.10
sqlmodel==0.0.22
pydantic~=2.10.3
passlib~=1.7.4
pydantic_core~=2.27.1
typing_extensions~=4.12.2
alembic~=1.14.0
python-dotenv~=1.0.1
bcrypt==4.2.1
python-jose==3.3.0
passlib==1.7.4
